requirements.txt
boto3==1.40.0
botocore==1.40.69
delta-spark==3.1.0
importlib_metadata==8.7.0
jmespath==1.0.1
kafka-python==2.0.3
py4j==0.10.9.7
pyspark==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.1
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
zipp==3.23.0
